refactor: context export 방식 변경
