+++ client/src/context/index.tsx
@@ -1,0 +1,13 @@
+import GeneralProvider from './GeneralProvider';
+import { CategoriesContext, ActiveCategoryIdContext, RenderCategoryIdContext } from './Category';
+import { OrderContext } from './Order';
+import { SelectedFoodContext } from './SelectedFood';
+
+export {
+  GeneralProvider,
+  CategoriesContext,
+  ActiveCategoryIdContext,
+  RenderCategoryIdContext,
+  OrderContext,
+  SelectedFoodContext,
+};
