+++ client/src/page/MainPage/Header/CategoryTab/index.tsx
@@ -1,6 +1,6 @@
 import { useContext } from 'react';
 import { TransperentButton } from '../../../../component';
-import { ActiveCategoryIdContext } from '../../../../context/CategoryContext';
+import { ActiveCategoryIdContext } from '../../../../context';
 import { click } from '../../../../util/pointerEvent';
 import styles from './CategoryTab.module.scss';
